requirements-refactored.txt
# Real-ESRGAN Ultimate - Refactored Version Requirements
# Install with: pip install -r requirements-refactored.txt

# Core Dependencies
torch>=1.7.0,<3.0.0
torchvision>=0.8.0
basicsr>=1.4.2
realesrgan>=0.3.0
gfpgan>=1.3.8

# UI Framework
gradio==5.23.3

# Image & Video Processing
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0
scipy>=1.5.0

# Utilities
python-dotenv>=0.19.0
requests>=2.25.0
tqdm>=4.50.0
pyyaml>=5.3.1

# Development & Testing (optional)
# Uncomment if developing or running tests
# pytest>=6.0.0
# pytest-cov>=2.12.0
# pytest-asyncio>=0.17.0

# Optional: Face enhancement
# Already included in gfpgan, but listed for clarity
# facexlib>=0.2.5

# Optional: Video codec support
# ffmpeg-python>=0.2.1
# This requires ffmpeg binary to be installed separately

# Note: Some packages may have additional system dependencies:
# - libzlib, libffi, liblzma (compression libraries)
# - libcuda, libnvrtc (for NVIDIA CUDA support)
# - For video: ffmpeg, libavcodec, libavformat
